+++ src/files/mod.rs
@@ -1,4 +1,5 @@
 pub mod append;
 pub mod build_paths;
+pub mod create;
 pub mod find;
 pub mod name;
