--- src/nots/mod.rs
@@ -1,2 +1,0 @@
-pub mod create;
-pub mod get;
